BigDecimal added! Logger started.

+++ src/main/java/com/techelevator/products/Candy.java
@@ -1,8 +1,10 @@
 package com.techelevator.products;
 
+import java.math.BigDecimal;
+
 public class Candy  extends Product {
 
-    public Candy(String name, double price) {
+    public Candy(String name, BigDecimal price) {
         super(name, price);
     }   // Candy constructor using the Product constructor
