+++ src/main/java/com/techelevator/products/Chip.java
@@ -1,8 +1,10 @@
 package com.techelevator.products;
 
+import java.math.BigDecimal;
+
 public class Chip  extends Product {
 
-    public Chip(String name, double price) {
+    public Chip(String name, BigDecimal price) {
         super(name, price);
     }   // Chip constructor using the Product constructor
