+++ src/main/java/com/techelevator/products/Gum.java
@@ -1,8 +1,10 @@
 package com.techelevator.products;
 
+import java.math.BigDecimal;
+
 public class Gum  extends Product {
 
-    public Gum(String name, double price) {
+    public Gum(String name, BigDecimal price) {
         super(name, price);
     }   // Gum constructor using the Product constructor
